Created mock glew system which rebinds all the the glew and gl functions to mock functions.
This consists of two parts:

The first of these is a dll called glmock which provides definitions for all the opengl calls with dll linkage and redirects them to calls to a mock object the interface of which it defines.

The second is a static library called glewmock. This provides a concrete implementation of the iglmock interface defined in the glmock dll and extends it to provide mock functions for all the remaining opengl and glew functions. I had to split this into two classes as the object file got too big due to the usage of extensive templates and COMDAT sections.

Added a memory leak detector to the unit tests, which then promptly discovered the leaks in the Transform tests which have now been fixed.